+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move RSA key checks to fix bug where bad RSA keys were causing issues at launch
 - Add cli command to export catalogue metadata to a Atlassian Confluence space
 - Integrate plugin tests (HICPluginTests, LoadModules.Extensions.Tests, Rdmp.Dicom.Tests) into main solution and test orchestration
+- FIX: Foreign Key visibility race condition in DB with AUTO_UPDATE_STATISTICS_ASYNC enabled
 
 
 ## [9.0.1] - 2025-07-31